Project 3/VEDANT/requirements.txt
langchain
weaviate-client
sentence-transformers
fastapi
uvicorn[standard]
streamlit
pandas
python-dotenv
tqdm
requests
faiss-cpu
transformers
torch